backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyyaml==6.0.1
python-multipart==0.0.6
panoramisk==1.4
aiohttp==3.9.1
